bug fix: Allow changing default ports
After adding nginx as a reverse proxy, 3 ports needs to be specified to
start Kimchi.
- port to list on (default 8000)
- ssl_port to enable SSL connections (default 8001)
- cherrypy_port, the cherrypy server port (default 8010)

The kimchi.conf was inconsistent to that current Kimchi logic.
So the user was not able to change the default ports without having
problems to start Kimchi.

Fix it by adding cherrypy_port to kimchi.conf file and adjusting the
kimchid script to get the correct variable values.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>