Fix i18n packaging scripts
There are some small problems in i18n packaging. Through commit bd98489
adds i18n.json.tmpl to Makefile.am, it forgets to add it to RPM spec.
This patch adds i18n.json.tmpl to RPM spec files.

Another problem is that plugin/sample/po/Makefile.in.in is linked to
po/Makefile.in.in, so it would install the sample.mo to the system by
default. This patch copies po/Makefile.in.in to plugin/sample/po, and
change the install/uninstall target to skip installing and uninstalling
the sample mo files.

The last problem is that plugins/sample/config.status is needed by
plugin/sample/po/Makefile.in.in to create the actual Makefile. In git
source code repo it is a soft link to top level config.status, but it
didn't gets distributed in the tar ball. This patch adds this soft link
file to the dist tar ball.

To test this patch, firstly run the following commands.
 make && make dist
Then
 ./autogen && make && make rpm

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>