Bug fix: Do not require login for simple resources
The subdirectories "css", "js", "libs" and "images" currently require
a login session to be accessed. Those files are simple resources, thus
they don't need to be protected.
One consequence of this behavior is that it is not possible to create a remote
template with the current security settings.

Remove the login session requirement to access the subdirectories "css",
"js", "libs" and "images".

Signed-off-by: Aline Manera <alinefm@br.ibm.com>
Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>