authorization: Add "access" elements to tabs.xml to describe user view
Kimchi has 2 user roles: "admin" with full control of Kimchi features
and "user" with limited access
To describe how each tab should be displayed to user, an "access"
element was added to tabs.xml indicating which view mode each role has.

<access role="..." mode="..."/>

The "mode" attribute values are:

- none: do not show the tab;
- admin: full instance access;
- read-only:  read-only access;
- byInstance: each resource will have its configuration sent by the
  backend ("access" parameter);

The user will only be able to manage the guests he/she is assigned for,
because that the guest tab has 'mode' == by-instance.
That way each VM will have a new parameter "access" indicating if user has
"full" or "read-only" access to this VM.

As a user can edit a guest, he/she may need to know which networks
and storage pools are configured, so set network and storage tab 'mode'
to read-only.

And as user should not perform any operation on host or templates, set
their 'mode' attributes to 'none'.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>