Get user groups correctly
Kimchi uses the Python API (module "grp") to get the groups which a user
belongs to. But that implementation is not correct, in some cases
some groups are left out.

For example, take a look at the following commands. Here's the Python
method of getting the user groups (user=vianac):

$ python -c "import grp; u = 'vianac'; print [ g.gr_name for g in grp.getgrall() if u in g.gr_mem ]"
['wheel', 'vianac', 'desktop_admin_r', 'aline']

And here's another method of getting the same groups, using a GNU/Linux
command:

$ id -Gn vianac
vianac wheel desktop_admin_r aline

Now, let's try the same thing with a different user (user=root):

$ python -c "import grp; u = 'root'; print [ g.gr_name for g in grp.getgrall() if u in g.gr_mem ]"
[]

$ id -Gn root
root

As shown above, the Python method doesn't always display the correct
results. As the command "id" is bundled in the GNU/Linux package
"coreutils", I'd say its output is the correct one.

Use the external command "id" to get the user groups instead of the
Python API.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>