Issue #405 - Fix enter hit in storage tab under guest edit window
If user goes to storage tab in edit guest window, whenever he focuses
a input boxes (any device name or path) and press ENTER, the default
behaviour is trigger a "click" event in the first window button. In this
case, the "Attach new storage" window is opened. This patch fixes this
problem, so, when user press ENTER, nothing happens.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>