Add function to list all pending debug reports
Only the user who has started the debug report creation was able to get
the view of a pending debug report. But while switching tabs, this
information was lost.
To fix this behavior and also let all users get the same view, the
pending debug reports must be listed while loading the report grid.
In order to do that, create a function to return all the pending debug
reports.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>