bug fix: Probe image file only during the template creation
We use guestfs to probe the OS name and version from an image file.
But it should be done only during the template creation. While lookup
for a template it is supposed to already has all the information in it.

To know when probing or not the image file or cdrom, VMTemplate has
the "scan" parameter: when set to true the OS name and version will be get
from the file; otherwise, use the already set values.

The current code was ignoring the "scan" value when probing the image
file. And as guestfs takes some time to do the scanning, the GET response also
delayed.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>