bug fix: Redirect user to the URL accessed prior to login
When a non-logged user tries to access a guest console:

https://localhost:8001/vnc_auto.html?port=64667&path=?token=dWJ1bnR1MTQuMDQ&encrypt=1

He/she will be asked to login and then he/she must be redirected back to the
same URL accessed prior to login (ie, the guest console).

To do that, we use the "next" query parameter. But commit 22eba438 wrongly
overrided this logic by only querying the last tab visited which was
redirecting user to the Kimchi main page instead of the guest console.
Fix it.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>