Modify spice_auto.html for Kimchi proposals
The sources were updated according to Kimchi configuration to get the
Javascript and CSS files from the installed spice-html5 package.
Kimchi is not using the default spice_auto.html because Kimchi uses
wss:// for all connections and it is only supported by recent versions of
spice-html5.
In addition to it, Kimchi points user to the right token on URL which
requires a change on spice_auto.html

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>