Update Kimchi to use the installed spice-html5
spice-html5 is present in almost all supported Linux distributions
(RHEL6.5, Fedora20 and Ubuntu 14.04). So in those cases we should not package
the spice-html5 code into Kimchi and instead of that uses the spice-html5 package
as dependency.

This patch change the build process to check the spice-html5 package
availability in order to decide to package or not the spice-html5 code.

spice-html5 package is not available for RHEL7 and openSUSE 13.1. So
for those distributions, you must run the following command to Kimchi
includes the spice-html5 into its package.

./autogen.sh --with-spice-html5

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>