snapshot: Create domain snapshots
A new command is added to create a new snapshot:

POST /vms/<vm-name>/snapshots {'name': '<snapshot-name>'}

It creates a new snapshot with the current state of the VM. Currently,
due to Kimchi not supporting paused states, snapshots can only be
created when the VM is shut off. The parameter 'name' is optional;
Kimchi will create a default value based on the current time if it's not
provided.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>