List as many pools' info as possible
If a pool happens to be busy at the time the pool.refresh() is
called, then libvirt will throw an error: "internal error:
pool 'default' has asynchronous jobs running." Currently, kimchi
rasise an exception and no pools' storage volumes are listed.
This patch returns a 0 for that particular pool, and allows
the user to continue to request information for other pools.

Without this patch, nothing is shown at all on the Storage tab
if one pool is busy.

In the future, we can check for VIR_ERR_RESOURCE_BUSY and
come up with a plan that works with the UI to display "Busy"
for a pool. But for now, this works around the issue.

v1 -> v2:
- Log exception message and error code

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>